backend/requirements.txt
flask
pytest
marshmallow
Flask-SQLAlchemy